Changed sleep alarms to use absolute times for interrupts. We divide the minute in the hour by 10 and set a new time dependant on the result (0-5).
Moved the time syncronisation so it triggers after the sensor reading
preventing the GPS sync period from adding to the time period between sensor
readings. The only change in the sensor period for that cycle should be the
clock correction.

Moved the check that decides if the time needs syncronation to be performed so
that it only happens once per hour, and added logic to postpone the time sync
by 25 hours if the periodic sync fails. On power up, a time sync will occur
within 1 hour, and be repeated every hour until it succeeds. It checks once
per hour to see if it needs to be done again. after 28 days it will trigger
another syncronisation. If that and subsequent syncronisations fails, then it
gets postponed by 25 hours before it attempts to do it again. That way we
should hopefully not kill the battery if the GPS fails to set the time for a
period.

Fixed error where we were sending the sd_error instead of ts_error for a
failed time sync