SD write code:
Simplify file writing, so files are all created in the root
directory.

Alter method for testing return values of SD operations as calling functions
from inside the if statement doesn't work, so we've altered it to store the
return value in a variable and then check the variable to identify the return
value of each operation. This has reduced the random failures of each
operation, and caused the SD.flag variable to be set correctly on error.

Main Loop:

Altered place where epoch is broken.
Reduced duration of LED flash on wake up.